Rename filters.py -> report_app_config.py
