+++ corehq/apps/app_manager/models/report_app_config.py
