+++ corehq/apps/app_manager/tests/test_report_config.py
@@ -21,7 +21,7 @@ from corehq.apps.app_manager.models import (
     ReportAppConfig,
     ReportModule,
 )
-from corehq.apps.app_manager.models.filters import (
+from corehq.apps.app_manager.models.report_app_config import (
     _filter_by_user_id,
     _get_auto_filter_function,
 )
